refactor(internal/surfer): inline gcloud generate from internal/librarian/gcloud
The Generate/GenerateConfig pair in internal/librarian/gcloud had exactly
one caller: internal/surfer. Move both into internal/surfer as unexported
generate/generateConfig and delete the now-empty internal/librarian/gcloud
package.

Keeping surfer-specific logic inside surfer straightens the dependency
graph and isolates the gcloud.yaml-driven path for eventual removal.
